Split scraping CF from being scraped
The Cloudflare analytics API exposes metrics in buckets of time, and
does not expose actual totals. In order to report accurate counters to
prometheus, we must update metrics on a timer loop, and expose them when
scraped.

Note that absolute values of these counters cannot be trusted as zero
is arbitrary.