Initial zeroing of HTTP requests by country
At startup, fetch a list of each country that has made an HTTP request
in the last 12 hours (the API maximum lookback is 15 hours).

This reduces the chances of missing data for low-traffic countries after
an exporter restart, which might not return any data in the first few
loops. Zeroes are seen by prometheus as a counter reset.