Use pagination to support large result sets
Poking the Cloudflare API reveals that the "limit" restricts the number
of buckets returned per zone. To paginate, we must detect that the
number of buckets we've received is equal to the limit, and send another
request with start time equal to the latest bucket time from the last
request. This can differ between zones, as the last minute finalized may
be different. Therefore, we now make a separate API request per zone, in
order to be able to specify different start times.

Passing in multiple "zones" objects doesn't work. The last one
specified will clobber all the others.

The semantics of double-counting / missing metrics are not changed, but
the implementation is changed a little to accommodate pagination. At
startup, the "last bucket seen" per zone is initialized to
scrapeInterval before the current time. If the exporter is restarted
quickly, Cloudflare will not be scrape until 1 scrape interval has
passed, and will only count buckets since that time.

When testing these queries (in code and in Graphiql), I noticed that
small requests (start time: 2-5 minutes ago, limit: 2) often return no
metrics. Increasing the limit, or the time window, returned metrics that
should have been returned the first time. This is not a race with
finalisation, and appears to be a Cloudflare bug. Using the hardcoded 5
minute grace time mitigates the risk of this, and by tracking the last
bucket seen we do not double count metrics.

Add a new development flag, "initial-scrape-immediately". This causes
the exporter to additionally scrape Cloudflare once at startup, to avoid
waiting for the first timer tick when testing changes.