Cap query window growth for quiet data sets
There was already mitigation in place for the case in which no data
points are returned. However, there was a scenario in which the same
data point would be returned across multiple successive overlapping
query windows. The last recorded timestamp would never be updated, and
the query window would grow until we receive API errors.

Note that the query windows overlap intentionally, as we cannot predict
when Cloudflare will "finalize" recent minutes. We track the last seen
timestamp per data set to avoid double counting.