Remove vector initialization
There is no obvious timestamp to set these metrics to. Using the current
time (either explicitly, or implicitly by prometheus on scrape), can
cause the next scrape loop to return metrics with timestamp before the
current time, causing prometheus to throw the errors "Error on ingesting
samples with different value but same timestamp" and "Error on ingesting
out-of-order samples".